work in progress
Projects + moved globe to layout + fixed socials + hehe
