+++ app/layout.tsx
@@ -4,6 +4,7 @@ import type { Metadata, Viewport } from "next";
 import { Roboto_Mono } from "next/font/google";
 import type { ReactNode } from "react";
 import "./globals.css";
+import Globe from "@/components/Globe";
 import { Socials } from "@/components/Socials";
 
 const Font = Roboto_Mono({ subsets: ["latin"] });
@@ -49,7 +50,7 @@ export const viewport: Viewport = {
 	width: "device-width",
 	initialScale: 1,
 	maximumScale: 1,
-	userScalable: false,
+	userScalable: true,
 };
 
 type Props = {
@@ -62,8 +63,14 @@ export default function RootLayout({ children }: Props) {
 			<body
 				className={`${Font.className} min-h-screen min-w-screen bg-[length:12rem] bg-topography bg-fixed bg-repeat`}
 			>
-				<div className="z-50 w-full min-w-screen p-4 md:hidden">
+				<div className="z-50 flex h-full w-full items-center justify-center p-4 md:fixed md:w-10 md:justify-start">
 					<Socials />
+				</div>
+				<div className="md:-right-[30%] -bottom-52 md:-bottom-32 -z-30 fixed h-full w-full transition-all duration-300">
+					<Globe />
+				</div>
+				<div className="-z-[5] fixed inset-0 bg-gradient-to-tr from-black via-black/60 to-transparent">
+					&nbsp;
 				</div>
 				{children}
 				<SpeedInsights />
