+++ package.json
@@ -35,8 +35,8 @@
 		"class-variance-authority": "^0.7.1",
 		"clsx": "^2.1.1",
 		"cobe": "^0.6.3",
-		"cookies-next": "^4.3.0",
-		"lucide-react": "^0.469.0",
+		"cookies-next": "^5.1.0",
+		"lucide-react": "^0.471.1",
 		"motion": "^11.18.0",
 		"next": "^15.1.4",
 		"nodemailer": "^6.9.16",
@@ -56,7 +56,7 @@
 	},
 	"devDependencies": {
 		"@biomejs/biome": "^1.9.4",
-		"@types/node": "^22.10.6",
+		"@types/node": "^22.10.7",
 		"@types/nodemailer": "^6.4.17",
 		"@types/react": "19.0.2",
 		"@types/react-dom": "19.0.2",
